fix: Stop small UI jumps when scrolling on network timelines (#1872)
Previous code returned a "large" page (~ 3x normal size) when refreshing
and the content was not in the page cache (i.e., content had to be
fetched by the remote mediator, as on the first load of a non-cached
timeline).

But it returned a normal size page when refreshing if the content was
already in the cache, as when a refresh was triggered because scrolling
caused the remote mediator to invalidate the paging source.

This appears to cause some small jumps when scrolling on non-cached
timelines.

The fix is to -- in the refresh case -- always try and return a large
page. If the page is not large then try and synthesize one using the
neighbour pages.