feat: Show whether a post is a reply in the post "header" (#1876)
Use the `status_info` view to display information if a post is a reply.

There are three possibilities:

1. Account is replying to itself, show "Continued thread"
2. Account is replying to another account we have cached information
for, show "Replied to `name`".
3. Account is replying to another account we don't have cached
information for, show "Replied".

The "`name`" portion of the content is shown in bold. Because this
increases the amount of information on the timeline set the size of the
header area to `status_text_small` (which also affects the "`name`
boosted" messages).

The header isn't shown in threads, where every status except the first
is a reply, as that would be too noisy.

Fixes #1813