feat: Display account pronouns, if present, in timelines etc
To do this, extract the pronouns from the account's "fields", if the
name of the field matches "pronouns" or similar in a variety of
languages. These are then persisted for display.

The seed lists of "pronouns" in various languages was extracted from
Fedilab and Moshidon, who have a similar feature.

Pronouns are displayed in a new `PronounsChip`. The chip has a low
colour background to indicate it is interactable and to low-key draw
attention to the content. The chip might not have enough width to
display the full pronoun content, so tapping the chip shows a toast
with the full content (or as much as a toast can display).

Pronouns are shown everywhere details for another account might appear:

- Status timelines
- Notifications
- When replying to an account
- Search results
- Autocomplete results (when creating a post)

To make room for this the regular account information is now always
displayed on two lines.

In timelines this is:

First line: Display name and status creation time / edit state
Second line: Handle, pronouns

Everywhere else it is:

First line: Display name, pronouns
Second line: Handle

Pronouns field values may be quite long. Rather than read them out in
Talkback on every selection, add an accessibility action that shows the
same Toast as tapping the chip.