fix: Allow for filters to be cleared on non-cached timelines (#1887)
When updating statuses based on user actions on non-cached timelines the
previous code was both (a) acting on the wrong status ID (should have
used the real ID, not the actionable ID), and (b) not calling the
provided `updater` function if the status was a reblog.

The net effect was that if a non-cached timeline, like a list, contained
a status that was both 1. a reblog, and 2. filtered, the "Clear filters"
button was displayed, but didn't do anything.

Fixes #1875