feat: Support push notifications on GoToSocial (#1888)
The `WebPushSubscription.id` property is inconsistently documented and
returned by different servers. In particular, documented as both a
string and a number https://github.com/mastodon/mastodon/issues/36609.

This breaks trying to register for push notifications on GoToSocial
servers. Since we don't use the property for anything just ignore it.

Fixes #1849