chore: Add lint checks for use/non-use of HTML entities in string resources (#1894)
Pachli code has a lot of:

```kotlin
val someString = HtmlCompat.fromHtml(
    context.getString(R.string.some_string, someValue),
    HtmlCompat.FROM_HTML_MODE_LEGACY
)
```

because it's the best way to get styled text from a string resource
while also interpolating values in to that string.

When doing this, the string resource **must** use entities in the HTML
tags. I.e., this:

```xml
<string name="some_string">This is &lt;b>bold&lt;/b></string>
```

Not

```xml
<string name="some_string">This is <b>bold</b></string>
```

If entities are not used then the HTML tags are ignored.

But, if you don't need to interpolate any values in to the string, and
use `getText()` then this is exactly reversed -- `getText()` will parse
the HTML in the string, and expects real angle brackets, not entities.

This is covered in
https://developer.android.com/guide/topics/resources/string-resource#StylingWithHTML.

This is easy to get wrong in the original language. It's even easier to
get wrong in translations, because of
https://github.com/WeblateOrg/weblate/issues/11995
where it's unclear to translators whether the original text is using
entities or not, and whether the text they've entered uses entities or
not.

Try and catch this with two new lint checks that use a simple heuristic.

`StringResourceAngleBracketDetector` looks for string resources that
contain literal `<` **and** appear to be referenced in a `getString`
call, and warns that this is probably incorrect.

`StringResourceEntityDetector` does the inverse, looking for string
resources that contain `&lt;` or `&gt;` **and** appear to be referenced
by `getText`, and warns that this is probably incorrect.