Use updateStatuses rather than upsertStatuses.
Upsert needs the account referenced in the status to be present, since
it will insert if the status is missing. If the account is not present
there's an FK constraint failure.  Since `update` does not insert it's
the better choice when opportunistically updating a status that may or
may not be cached.