feat: Show quote posts (#1919)
For now, quotes are shown below everything else in the status (polls,
attachments, etc).

Android does not like recursive views, so a new `QuotedStatusView` is
introduced to display the content. This supports the "internal"
interaction controls on a status; clicking through filters, showing
content hidden by a content warning, expanding content that is too long,
voting, attachments, etc.

It does not have its own controls for replying, etc. The user has to
click through to view the quoted status directly for those.

An accessibility action is added to allow the user to do that.

Quotes are shown on a darker background to differentiate them from
the rest of the content. This presentation is now also used for preview
cards.

In the database this is implemented as a new `TimelineStatusWithQuote`
which joins the parent status and the status it is quoting (if any).

In the UI this is implemented as a new `StatusItemViewData` to hold the
parent status and the status it is quoting (if any). `IStatusViewData`
is still used for each of the two statuses that make up the quoting
pair.

This allows all the code that operates on an `IStatusViewData` to be
used for the parent post and the quote.

When fetching statuses the `RemoteMediator` implementations write out
any quoted statuses as well as any accounts referenced by those statuses
so everything can be put back together.

This has moved some data around. For example, in `ConversationViewData`
the `attachmentDisplayAction` used to be a top-level property in that
class. Now it is -- properly -- a property of the viewdata specific to
that status, since the visibility of the attachments in a status that
contains a quote can be controlled for both the parent status and the
quoted status.

Content filters are applied to quoted statuses, account filters are not,
on the assumption that someone you trust quoting someone you might
otherwise filter is probably safe, and a good way to be introduced to
accounts you might want to follow.

As part of this work:

- `NetworkTimelineRepository` was making two database calls per status
on every load. That's been pushed down to `NetworkTimelinePagingSource`
where it can be two database calls per page of statuses instead.

- To keep queries manageable, `TimelineStatusWithAccount` is now a
view on the database.