fix: Correctly prune database cache (#1932)
The queries used to prune the cache were returning empty results if
notifications with null status IDs existed (which is highly likely). As
a result of that the database wasn't being correctly pruned.

Fix that by updating the query.

In addition, add a developer tool entry to prune the cache, and add it
to the new "Database" tab on the "About" screen.