refactor: Prune cache once at start, retain more referenced statuses (#1939)
Previous cache pruning code only kept statuses if they were referenced
directly by `TimelineStatusEntity`, `NotificationEntity`, or
`ConversationEntity`.

Extend this to include:

1. Statuses referenced by `DraftEntity`
2. Statuses referenced by any of these statuses in their
`reblogServerId`, `inReplyToId`, or `quoteServerId` properties
(recursively).

Put the logic for determining the list of referenced status IDs in a new
view, `ReferencedStatusId`. Use this to decide whether a status should
be retained in the different `cleanup...` methods.

Wipe the cache completely immediately after upgrade to ensure the user
isn't stuck with an unpruned cache for ~ 24 hours.

The new query requires an up-to-date SQLite, so make sure the
BundleSQLiteDriver is used in device tests.