fix: Load images from sites that use LetsEncrypt certificates
Previous code loaded embedded LetsEncrypt certificates on Android 6 /
API 23 for the OkHttpClient used for API requests, but the client used
by Glide was not affected, so requests to load images from those sites
would fail due to SSL errors.

Fix that, and have Glide use the same OkHttpClient as the rest of the
app. That should use fewer resources, share the cache between the two,
and pick up the special SSL certificate customisation.