refactor: Update translation and menu handling in SFragment (#1948)
SFragment defined the methods `onTranslate` and `onTranslateUndo`, when
those should have been in `StatusActionListener`, as these are actions
the user may take on a status.

The "..." (more) menu handling code in SFragment had to have special
cases for conversations (direct messages). Split that out and provide
methods `ConversationsFragment` can call through to customise the menu
and provide conversation-specific handlers for.