feat: Support removing quotes from posts (#1949)
For this to work we need to know if the quote was authored by the user.
This is tracked in the new `isUsersStatus` property, and for this to
work `StatusViewData.from` needs to know the user's server ID. To make
this straightforward pass down the entire `PachliAccount`, not just the
ID, adjusting the call sites as appropriate.

Implement the "revoke quote" API, and access via `StatusRepository`.

To revoke a quote `QuotedStatusView` needs the ID of the parent status,
so provide that in a new `QuotedStatusViewData`.

Adjust some view parameters to get everything looking OK. In particular,
move the padding into `status_content_quoted`, and adjust the bottom
padding depending on whether the "Remove quote" button is present.