feat: Provide a user preference to control pronoun display
Add a user preference to control pronoun display, with options:

- Everywhere
- When writing a post
- Nowhere

The default is "When writing a post", as this balances showing the
information where it is most relevant without overloading the user
with information with the default display options.

Noticed the displays for "Blocked accounts", "Muted accounts", and
"Follow requests" didn't have the chip to display the pronouns so
added that. Redesigned "Follow requests" to move the reject/accept
buttons to the bottom of the item and display as icon+text to make
room for the pronoun chip.