feat: Show an account preference to set the default quote policy (#1959)
Read the account preference when logging in and save to a new property
in `AccountEntity`.

If not provided (e.g., the server does not support it) the default is
"NOBODY", as that's the expected behaviour.

Show a preference for this in `AccountPreferencesFragment`, with a check
to see if the server supports setting this property. If it doesn't the
preference is disabled.