feat: Quote posts (#1960)
Add support for quoting posts.

Record whether the server supports quote posts in `StatusDisplayOptions`
new `canQuote` property.

Add `StatusActionListener.onQuote()`, called when the user wants to
quote a status.

Update `StatusControlView`:

- Possibly show "quote" related menu items when clicking the "reblog"
  button.
- Call `listener.onQuote()` as appropriate.

Update `ComposeOptions`. At the moment a status referenced in
`ComposeOptions` is either a reply or a quote, so wrap that in a
sealed class so the UI for showing a reply can be repurposed to also
show a quote.

When composing:
- Add a new bottomsheet/menu to allow the user to set the quote policy.
- Display the status being quoted.
- Disable "DIRECT" visibility if quoting.
- Show the content being quoted at the top.

Include the quote policy and quoted content when editing a status, or
editing a scheduled status.

If set, include the ID of the quoted status, and the user's quote
policy whenever a status is sent.

Update `DraftEntity` and related infrastructure to support saving and
restoring (a) the user's quote policy, and (b) the status the user may
be quoting.

Add an accessibility action for quoting a post.