feat: Add a "..." menu to open the quoted post
If a quoted post is just attachments (e.g., just a single image) you
can't open the quoted post -- tapping on the header area takes you to
the account, tapping on the attachment opens it for viewing.

Add an "Open quoted post" item to the "..." menu, displayed if there
is a quoted post. If tapped it opens the thread for the quoted post.

Reported by @jandi@mastodon.social.