fix: Don't crash when showing suspending dialogs
Dialogs must be shown on the main (UI) thread. Enforce this in the
suspend dialog extensions by explicitly switching to `Dispatchers.Main`.

Fixes a crash; EnableAllNotificationsUseCase runs on `Dispatchers.IO`
(because it interacts with `SharedPreferences` which can perform IO)
and can show a dialog. The dialog was being displayed on an IO thread
and crashed with:

Can't create handler inside thread Thread[DefaultDispatcher-worker-9,5,main] that has not called Looper.prepare()

Fixes #1986