feat: Allow multiple simultaneous "compose" tasks
Send each `ComposeActivity` intent with `FLAG_ACTIVITY_NEW_DOCUMENT`
and `FLAG_ACTIVITY_MULTIPLE_TASK`.

Each `ComposeActivity` started with a unique intent will be placed in
a different task, allowing the user to swipe between them on the
"Recents" screen.

This also means that opening a notification (which navigates back to
the notifications tab) doesn't close any instances of `ComposeActivity`.

Fixes #1555