fix: Prevent OOM, actually prune the database (#2012)
`onPostMigrate(db: SupportSQLiteDatabase)` **is not called** if Room is
configured with a driver (see
https://developer.android.com/reference/androidx/room/migration/AutoMigrationSpec#onPostMigrate(androidx.sqlite.db.SupportSQLiteDatabase)).

So the previous code wasn't running, the database wasn't being pruned,
and users were seeing OOM crashes.

Fix this by overriding the version that takes `SQLiteConnection`. This
requires a new database version because users may already have installed
a version with the no-op migration.