feat: Show compose action in task label in "Recents" view
Now `ComposeActivity` appears as its own task in the "Recents" view
the label can be set differently to distinguish between the different
tasks.

Do this, with one of either:

- Writing post
- Replying to (name)
- Quoting (name)

This always affects Talkback. It doesn't affect the UI on some devices
because of https://issuetracker.google.com/issues/340133008.

Suggested by @jonathan859@someplace.social