fix: Show animated avatars in preview card bylines
If the avatar in a preview card byline was animated the animation wasn't
being started, so the avatar didn't appear.

Fix that. Write two new `CustomTarget` implementations that can load
bitmaps and drawables into a `TextView` drawable, and select between
them based on whether animation is enabled and the avatar is being
loaded as a bitmap or a drawable.