ci: Prefix Pachli Current releases with a date/time stamp
Pachli Current releases are pushed to `pachli-android-current`, which
doesn't have the code or commits, so it seems as though GitHub assigns
each release to the same commit then orders by name. This puts the
releases out of order making them difficult to navigate.

Try and mitigate this by prefixing the release name with a date/time
stamp.